client(Setup): moved and refactored features into the app folder
